Implement non-routables feature for Amazon route scanner
Add support for scanning and optimally placing non-routable packages in Amazon routes.
These are packages without sequence numbers that need to be delivered on the route.

Key changes:
1. Database schema (v10): Add isNonRoutable field to RoutedStopEntity, change sequenceOrder to Float to support fractional sequence numbers (e.g., 22.5)

2. Core optimization logic: New NonRoutableOptimizer class that calculates optimal placement between stops based on minimal route detour using Haversine distance

3. Scanner UI: Add mode toggle (Regular Stops / Non-Routables) with ScanMode enum and updated state management

4. Route player: Auto-optimize non-routable placement when entering play mode

5. Type updates across the codebase:
   - sequenceOrder: Int -> Float in RoutedStopEntity and related data classes
   - Updated all related queries and operations to handle float sequences

6. Integration in repository layer:
   - optimizeNonRoutableStopPlacement() - re-optimize stops using detour algorithm
   - updateStops() wrapper for batch updates
   - createStopsFromScannedAddresses() handles isNonRoutable flag

Non-routables are initially added at the end during scanning, then re-optimized
when the route is saved to the database based on actual coordinates.

+++ app/src/main/java/com/isaacshub/app/routehelper/ui/player/PackageScanScreen.kt
@@ -70,7 +70,7 @@ import java.util.concurrent.Executors
 data class ScannedPackage(
     val trackingNumber: String,
     val addressLabel: String,
-    val sequenceNumber: Int? = null,
+    val sequenceNumber: Float? = null,
     val sections: List<String> = emptyList(),
     val routedStopId: Long? = null,
     /** True if this package is for an unknown address on a known street */
